--- a/crates/accelerate/src/euler_one_qubit_decomposer.rs
+++ b/crates/accelerate/src/euler_one_qubit_decomposer.rs
@@ -1087,7 +1087,7 @@ pub(crate) fn optimize_1q_gates_decomposition(
             None => raw_run.len() as f64,
         };
         let qubit: PhysicalQubit = if let NodeType::Operation(inst) = &dag.dag[raw_run[0]] {
-            dag.get_qubits(inst.qubits)[0].into()
+            PhysicalQubit::new(dag.get_qubits(inst.qubits)[0].0)
         } else {
             unreachable!("nodes in runs will always be op nodes")
         };