--- a/crates/accelerate/src/nlayout.rs
+++ b/crates/accelerate/src/nlayout.rs
@@ -14,6 +14,7 @@ use pyo3::prelude::*;
 use pyo3::types::PyList;
 
 use hashbrown::HashMap;
+use qiskit_circuit::Qubit;
 
 /// A newtype for the different categories of qubits used within layouts.  This is to enforce
 /// significantly more type safety when dealing with mixtures of physical and virtual qubits, as we
@@ -24,7 +25,7 @@ use hashbrown::HashMap;
 macro_rules! qubit_newtype {
     ($id: ident) => {
         #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-        pub struct $id(u32);
+        pub struct $id(pub u32);
 
         impl $id {
             #[inline]
@@ -72,6 +73,25 @@ impl PhysicalQubit {
         layout.phys_to_virt[self.index()]
     }
 }
+
+/// This is only safe in the context of a physical circuit during transpilation
+/// after the qubit indices of the circuit/dag circuit refer to the physical
+/// qubits (once we've run a layout pass and applied it).
+impl From<Qubit> for PhysicalQubit {
+    fn from(s: Qubit) -> PhysicalQubit {
+        PhysicalQubit::new(s.0)
+    }
+}
+
+/// This is only safe in the context of a physical circuit during transpilation
+/// after the qubit indices of the circuit/dag circuit refer to the physical
+/// qubits (once we've run a layout pass and applied it).
+impl From<&Qubit> for PhysicalQubit {
+    fn from(s: &Qubit) -> PhysicalQubit {
+        PhysicalQubit::new(s.0)
+    }
+}
+
 qubit_newtype!(VirtualQubit);
 impl VirtualQubit {
     /// Get the physical qubit that currently corresponds to this index of virtual qubit in the