… crates/accelerate/src/two_qubit_decompose.rs
… crates/circuit/src/dag_circuit.rs
… qiskit/transpiler/passes/optimization/optimize_1q_decomposition.py
… releasenotes/notes/optimize-1q-gates-decomposition-ce111961b6782ee0.yaml