@mtreinish 2024-08-23
Now that this is ready for merging I ran two benchmark runs for this pass against main:

```
Benchmarks that have improved:

| Change   | Before [d3040a0b] <main>            | After [2c0aad55] <use-dag-1q-decompose>   |   Ratio | Benchmark (Parameter)                                                                                         |
|----------|-------------------------------------|-------------------------------------------|---------|---------------------------------------------------------------------------------------------------------------|
| -        | 5.50±0.03ms                         | 1.24±0ms                                  |    0.23 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(5, 1024, ['rx', 'ry', 'rz', 'r', 'rxx', 'id'])  |
| -        | 14.0±0.08ms                         | 2.95±0.02ms                               |    0.21 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(14, 1024, ['rx', 'ry', 'rz', 'r', 'rxx', 'id']) |
| -        | 24.3±0.2ms                          | 4.11±0.02ms                               |    0.17 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(20, 1024, ['rx', 'ry', 'rz', 'r', 'rxx', 'id']) |
| -        | 4.12±0.03ms                         | 572±1μs                                   |    0.14 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(5, 1024, ['u', 'cx', 'id'])                     |
| -        | 15.9±0.06ms                         | 2.01±0.01ms                               |    0.13 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(20, 1024, ['u', 'cx', 'id'])                    |
| -        | 5.99±0.02ms                         | 768±4μs                                   |    0.13 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(5, 1024, ['rz', 'x', 'sx', 'cx', 'id'])         |
| -        | 15.7±0.2ms                          | 1.92±0.03ms                               |    0.12 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(14, 1024, ['rz', 'x', 'sx', 'cx', 'id'])        |
| -        | 11.8±0.05ms                         | 1.44±0.01ms                               |    0.12 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(14, 1024, ['u', 'cx', 'id'])                    |
| -        | 28.1±0.2ms                          | 2.63±0.01ms                               |    0.09 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(20, 1024, ['rz', 'x', 'sx', 'cx', 'id'])        |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```

and against 1.2.0 (to discount any potential performance change from https://github.com/Qiskit/qiskit/pull/12550:
```
Benchmarks that have improved:

| Change   | Before [7e2e835e] <1.2.0^0>   | After [2c0aad55] <use-dag-1q-decompose>   |   Ratio | Benchmark (Parameter)                                                                                         |
|----------|-------------------------------|-------------------------------------------|---------|---------------------------------------------------------------------------------------------------------------|
| -        | 5.53±0.02ms                   | 1.23±0.01ms                               |    0.22 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(5, 1024, ['rx', 'ry', 'rz', 'r', 'rxx', 'id'])  |
| -        | 14.0±0.1ms                    | 2.98±0.03ms                               |    0.21 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(14, 1024, ['rx', 'ry', 'rz', 'r', 'rxx', 'id']) |
| -        | 25.7±0.1ms                    | 4.08±0.03ms                               |    0.16 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(20, 1024, ['rx', 'ry', 'rz', 'r', 'rxx', 'id']) |
| -        | 4.14±0.02ms                   | 584±10μs                                  |    0.14 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(5, 1024, ['u', 'cx', 'id'])                     |
| -        | 5.93±0.02ms                   | 778±10μs                                  |    0.13 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(5, 1024, ['rz', 'x', 'sx', 'cx', 'id'])         |
| -        | 15.8±0.09ms                   | 1.95±0.03ms                               |    0.12 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(14, 1024, ['rz', 'x', 'sx', 'cx', 'id'])        |
| -        | 12.1±0.09ms                   | 1.49±0.02ms                               |    0.12 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(14, 1024, ['u', 'cx', 'id'])                    |
| -        | 16.5±0.1ms                    | 2.06±0.04ms                               |    0.12 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(20, 1024, ['u', 'cx', 'id'])                    |
| -        | 30.5±0.2ms                    | 2.61±0.05ms                               |    0.09 | passes.MultipleBasisPassBenchmarks.time_optimize_1q_decompose(20, 1024, ['rz', 'x', 'sx', 'cx', 'id'])        |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```